+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support API Gateway Request Validation policies
 - Support HCX Enterprise Add-on for `OCVP` service
+- Support for Dynamic Limits
 
 ## 4.27.0 (May 19, 2021)